+++ assets/v030/main.js
@@ -2,7 +2,7 @@
  * @Author: laixi
  * @Date:   2017-03-14 11:34:34
  * @Last Modified by:   Xavier Yin
- * @Last Modified time: 2017-03-27 00:00:03
+ * @Last Modified time: 2017-04-24 11:12:22
  */
 import _ from 'underscore';
 import Attributes from './attributes';
@@ -22,6 +22,9 @@ _.extend(Backbone, Events);
 _.each(['Model', 'View', 'Collection', 'Router', 'History'], function(klass) {
   _.extend(Backbone[klass].prototype, Events);
 });
+
+// 扩展 Model 实例方法，增加 allChanged 和 anyChanged 方法。
+_.extend(Model.prototype, _.pick(Attributes, 'allChanged', 'anyChanged'));
 
 // this is invalid in es6
 module.exports = Backbone;